requirements-prod.txt
# Production Requirements
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data Validation
pydantic[email]==2.5.0
validators==0.22.0

# Caching
redis==5.0.1

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Utilities
python-slugify==8.0.1
python-dotenv==1.0.0

# Monitoring & Logging
sentry-sdk[fastapi]==1.38.0
structlog==23.2.0

# Rate Limiting
slowapi==0.1.9

# Configuration
pydantic-settings==2.1.0

# Production Server
gunicorn==21.2.0
